Backend/requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
nltk==3.8.1
python-dotenv==1.0.0 